+++ crates/ribosome/data/keys.rs
@@ -3,6 +3,8 @@
 use std::fmt;
 
 /// A key for reference sequences, combining a reference ID and a compound type.
+///
+/// Neither field will contain tabs.
 #[derive(Hash, PartialEq, Eq, Debug, Clone)]
 pub struct RefKey {
     /// The reference ID of the reference sequence (e.g., `ANHUI01`,
@@ -15,6 +17,10 @@ pub struct RefKey {
 impl RefKey {
     /// Creates a new [`RefKey`] by combining a `reference_id` (e.g., `ANHUI01`)
     /// and `compound_type` (e.g., `A_HA_H7`).
+    ///
+    /// ## Validity
+    ///
+    /// `reference_id` and `compound_type` cannot contain any tabs.
     pub fn new(reference_id: impl Into<String>, compound_type: impl Into<String>) -> Self {
         Self {
             reference_id:  reference_id.into(),
@@ -26,11 +32,34 @@ impl RefKey {
     /// `reference_id|compound_type`.
     ///
     /// Any additional pipe-delimited fields are ignored.
-    pub fn parse(name: &str) -> Option<Self> {
+    ///
+    /// ## Errors
+    ///
+    /// The header must successfully parse given the above format, and no tabs
+    /// can be present in the `reference_id` or `compound_type`. Context
+    /// including the `name` is included.
+    pub fn parse(name: &str) -> std::io::Result<Self> {
         let mut parts = name.split('|');
-        let reference_id = parts.next()?;
-        let compound_type = parts.next()?;
-        Some(Self::new(reference_id, compound_type))
+
+        let (Some(reference_id), Some(compound_type)) = (parts.next(), parts.next()) else {
+            return Err(std::io::Error::other(format!(
+                "Expected format reference_id|compound_type, found: {name}"
+            )));
+        };
+
+        if reference_id.contains('\t') {
+            return Err(std::io::Error::other(format!(
+                "The reference ID cannot contain a tab. Found: {name}"
+            )));
+        }
+
+        if compound_type.contains('\t') {
+            return Err(std::io::Error::other(format!(
+                "The compound type cannot contain a tab. Found: {name}"
+            )));
+        }
+
+        Ok(Self::new(reference_id, compound_type))
     }
 }
 
@@ -83,7 +112,24 @@ mod tests {
     }
 
     #[test]
+    fn test_ref_key_parse_extra_field() {
+        let key = RefKey::parse("A_HA|H3N2|more\tmore|more").unwrap();
+        assert_eq!(key.reference_id, "A_HA");
+        assert_eq!(key.compound_type, "H3N2");
+    }
+
+    #[test]
     fn test_ref_key_parse_missing_pipe() {
-        assert!(RefKey::parse("no_pipe").is_none());
+        assert!(RefKey::parse("no_pipe").is_err());
+    }
+
+    #[test]
+    fn test_ref_key_parse_ref_id_tab() {
+        assert!(RefKey::parse("A\tHA|H3N2").is_err());
+    }
+
+    #[test]
+    fn test_ref_key_parse_ctype_tab() {
+        assert!(RefKey::parse("A_HA|H3\tN2").is_err());
     }
 }